removing discovery backfill FF and discovery SQS queue (#1183)
