+++ .env.example
@@ -10,8 +10,6 @@ SQS_BACKFILL_QUEUE_URL=http://127.0.0.1:4566/000000000000/backfill
 SQS_BACKFILL_QUEUE_REGION=us-west-1
 SQS_PUSH_QUEUE_URL=http://127.0.0.1:4566/000000000000/push
 SQS_PUSH_QUEUE_REGION=us-west-1
-SQS_DISCOVERY_QUEUE_URL=http://127.0.0.1:4566/000000000000/discovery
-SQS_DISCOVERY_QUEUE_REGION=us-west-1
 SQS_DEPLOYMENT_QUEUE_URL=http://127.0.0.1:4566/000000000000/deployment
 SQS_DEPLOYMENT_QUEUE_REGION=us-west-1
 SQS_BRANCH_QUEUE_URL=http://127.0.0.1:4566/000000000000/branch
