+++ .env.test
@@ -10,8 +10,6 @@ SQS_BACKFILL_QUEUE_URL=http://127.0.0.1:4566/000000000000/test-backfill
 SQS_BACKFILL_QUEUE_REGION=us-west-1
 SQS_PUSH_QUEUE_URL=http://127.0.0.1:4566/000000000000/test-push
 SQS_PUSH_QUEUE_REGION=us-west-1
-SQS_DISCOVERY_QUEUE_URL=http://127.0.0.1:4566/000000000000/test-discovery
-SQS_DISCOVERY_QUEUE_REGION=us-west-1
 SQS_DEPLOYMENT_QUEUE_URL=http://127.0.0.1:4566/000000000000/test-deployment
 SQS_DEPLOYMENT_QUEUE_REGION=us-west-1
 SQS_TEST_QUEUE_URL=http://127.0.0.1:4566/000000000000/test-sqs-client
