+++ .localstack/sqs.sh
@@ -3,7 +3,6 @@
 # Development queues
 awslocal sqs create-queue --queue-name backfill
 awslocal sqs create-queue --queue-name push
-awslocal sqs create-queue --queue-name discovery
 awslocal sqs create-queue --queue-name deployment
 awslocal sqs create-queue --queue-name branch
 
@@ -11,6 +10,5 @@ awslocal sqs create-queue --queue-name branch
 awslocal sqs create-queue --queue-name test-sqs-client
 awslocal sqs create-queue --queue-name test-backfill
 awslocal sqs create-queue --queue-name test-push
-awslocal sqs create-queue --queue-name test-discovery
 awslocal sqs create-queue --queue-name test-deployment
 awslocal sqs create-queue --queue-name test-branch
