+++ docker-compose.yml
@@ -52,6 +52,5 @@ services:
       SQS_PUSH_QUEUE_URL: http://localstack:4566/000000000000/push
       SQS_DEPLOYMENT_QUEUE_URL: http://localstack:4566/000000000000/deployment
       SQS_BRANCH_QUEUE_URL: http://localstack:4566/000000000000/branch
-      SQS_DISCOVERY_QUEUE_URL: http://localstack:4566/000000000000/discovery
       SQS_TEST_QUEUE_URL: http://localstack:4566/000000000000/test-sqs-client
       REDISX_CACHE_HOST: redis
