+++ src/config/env.ts
@@ -16,8 +16,6 @@ const requiredEnvVars = [
 	"SQS_BACKFILL_QUEUE_REGION",
 	"SQS_PUSH_QUEUE_URL",
 	"SQS_PUSH_QUEUE_REGION",
-	"SQS_DISCOVERY_QUEUE_URL",
-	"SQS_DISCOVERY_QUEUE_REGION",
 	"SQS_DEPLOYMENT_QUEUE_URL",
 	"SQS_DEPLOYMENT_QUEUE_REGION",
 	"SQS_BRANCH_QUEUE_URL",
@@ -68,8 +66,6 @@ export interface EnvVars {
 	SQS_BACKFILL_QUEUE_REGION: string;
 	SQS_PUSH_QUEUE_URL: string;
 	SQS_PUSH_QUEUE_REGION: string;
-	SQS_DISCOVERY_QUEUE_URL: string;
-	SQS_DISCOVERY_QUEUE_REGION: string;
 	SQS_DEPLOYMENT_QUEUE_URL: string;
 	SQS_DEPLOYMENT_QUEUE_REGION: string;
 	SQS_BRANCH_QUEUE_URL: string;
