+++ src/sqs/queues.ts
@@ -3,7 +3,6 @@ import { SqsQueue } from "./sqs";
 import { BackfillMessagePayload, backfillQueueMessageHandler } from "./backfill";
 import { pushQueueMessageHandler, PushQueueMessagePayload } from "./push";
 import { jiraAndGitHubErrorsHandler, webhookMetricWrapper } from "./error-handlers";
-import { DiscoveryMessagePayload, discoveryQueueMessageHandler } from "./discovery";
 import { DeploymentMessagePayload, deploymentQueueMessageHandler } from "./deployment";
 import { BranchMessagePayload, branchQueueMessageHandler } from "./branch";
 import { getLogger } from "config/logger";
@@ -34,18 +33,6 @@ export const sqsQueues = {
 		maxAttempts: 5
 	}, pushQueueMessageHandler, webhookMetricWrapper(jiraAndGitHubErrorsHandler, "push")),
 
-	discovery: new SqsQueue<DiscoveryMessagePayload>({
-		queueName: "discovery",
-		queueUrl: envVars.SQS_DISCOVERY_QUEUE_URL,
-		queueRegion: envVars.SQS_DISCOVERY_QUEUE_REGION,
-		longPollingIntervalSec: LONG_POLLING_INTERVAL_SEC,
-		timeoutSec: 10 * 60,
-		maxAttempts: 3
-	},
-	discoveryQueueMessageHandler,
-	jiraAndGitHubErrorsHandler
-	),
-
 	deployment: new SqsQueue<DeploymentMessagePayload>({
 		queueName: "deployment",
 		queueUrl: envVars.SQS_DEPLOYMENT_QUEUE_URL,
@@ -74,7 +61,6 @@ export const sqsQueues = {
 		logger.info("Starting queues");
 		sqsQueues.backfill.start();
 		sqsQueues.push.start();
-		sqsQueues.discovery.start();
 		sqsQueues.deployment.start();
 		sqsQueues.branch.start();
 		logger.info("All queues started");
@@ -85,7 +71,6 @@ export const sqsQueues = {
 		await Promise.all([
 			sqsQueues.backfill.stop(),
 			sqsQueues.push.stop(),
-			sqsQueues.discovery.stop(),
 			sqsQueues.deployment.stop(),
 			sqsQueues.branch.stop()
 		]);
